rename aggregate source file
